add: MyBatis Migrations

+++ src/main/java/com/example/demo/domain/Post/PostDto.java
@@ -10,7 +10,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
  */
 public class PostDto {
   @Schema(description = "投稿 ID", example = "1", requiredMode = Schema.RequiredMode.REQUIRED)
-  private Integer id;
+  private Long id;
   @Schema(description = "ユーザー ID", example = "example", requiredMode = Schema.RequiredMode.REQUIRED)
   private String userId;
   @Schema(description = "本文", example = "がんばります。", requiredMode = Schema.RequiredMode.REQUIRED)
@@ -18,14 +18,14 @@ public class PostDto {
   @Schema(description = "投稿日時", example = "2025-03-10T08:14:12.451+00:00", requiredMode = Schema.RequiredMode.REQUIRED)
   private Date createdAt;
 
-  public PostDto(Integer id, String userId, String content, Date createdAt) {
+  public PostDto(Long id, String userId, String content, Date createdAt) {
     this.id = id;
     this.userId = userId;
     this.content = content;
     this.createdAt = createdAt;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
